Add first brutally rudimentary Icosoku solver
Uses only the constraint that the sum at each vertex is less than the
vertex value, not equal to it.  Therefore it is very slow.

Numbering of the vertices is as follows: trigonometric increasing when
viewed from above, 0 at the top, down-left 1, down-right 6, bottom 11.
Tiles are read (once placed in a face) trigonometrically, beginning at
the "upper-leftmost" vertex.